split code, sharding-0.6.4
